requirements.txt
streamlit
opencv-python-headless
numpy
cvzone
ultralytics
opencv-python
streamlit
opencv-python
pytesseract
pillow